refactor(acceptance): extract scenario 28 helper group into sibling file (warren-acf2)
